Moved, renamed and added test for StringLengthComparator
