+++ src/main/java/net/sf/jabref/logic/formatter/casechanger/CaseKeeper.java
@@ -17,9 +17,9 @@ package net.sf.jabref.logic.formatter.casechanger;
 
 import java.util.Arrays;
 
-import net.sf.jabref.importer.fetcher.LengthComparator;
 import net.sf.jabref.logic.formatter.Formatter;
 import net.sf.jabref.logic.l10n.Localization;
+import net.sf.jabref.logic.util.strings.StringLengthComparator;
 
 public class CaseKeeper implements Formatter {
 
@@ -27,7 +27,7 @@ public class CaseKeeper implements Formatter {
         if (text == null) {
             return null;
         }
-        Arrays.sort(listOfWords, new LengthComparator());
+        Arrays.sort(listOfWords, new StringLengthComparator());
         // For each word in the list
         for (String listOfWord : listOfWords) {
             // Add {} if the character before is a space, -, /, (, [, ", or } or if it is at the start of the string but not if it is followed by a }
